+++ src/classes/features/ChillRoom.ts
@@ -27,6 +27,7 @@ import {
 } from "isaacscript-common";
 import { ChallengeCustom } from "../../enums/ChallengeCustom";
 import { mod } from "../../mod";
+import { preventSaveAndQuit } from "../../utils";
 
 const FONT = fonts.droid;
 
@@ -42,8 +43,10 @@ export class ChillRoom extends ModFeature {
     FONT.DrawString(text, 0, 33, KColorDefault, screenBottomRightPos.X, true);
   }
 
-  @CallbackCustom(ModCallbackCustom.POST_GAME_STARTED_REORDERED, undefined)
-  postGameStartedReordered(): void {
+  @CallbackCustom(ModCallbackCustom.POST_GAME_STARTED_REORDERED, false)
+  postGameStartedReorderedFalse(): void {
+    preventSaveAndQuit();
+
     const hud = game.GetHUD();
     hud.SetVisible(false);