+++ src/classes/features/Timer.ts
@@ -1,14 +1,15 @@
 import { ModCallback } from "isaac-typescript-definitions";
-import { Callback, ModFeature } from "isaacscript-common";
+import { Callback } from "isaacscript-common";
+import { isTimerEnabled } from "../../config";
 import { getModifiedBossID } from "../../enums/BossIDCustom";
 import { TimerType } from "../../enums/TimerType";
 import { timerDraw } from "../../timer";
 import { getNumSecondsForBossObjective } from "../../types/Objective";
+import { RandomizerModFeature } from "../RandomizerModFeature";
 import { getSecondsSinceLastDamage } from "./BossNoHitObjectiveDetection";
 import { getPlaythroughSecondsElapsed } from "./StatsTracker";
 
-// TODO: RANDOMIZER MOD FEATURE
-export class Timer extends ModFeature {
+export class Timer extends RandomizerModFeature {
   @Callback(ModCallback.POST_RENDER)
   postRender(): void {
     this.drawMainTimer();
@@ -16,10 +17,10 @@ export class Timer extends ModFeature {
   }
 
   drawMainTimer(): void {
-    // if (isTimerEnabled()) {
-    const seconds = getPlaythroughSecondsElapsed();
-    timerDraw(TimerType.MAIN, seconds);
-    // }
+    if (isTimerEnabled()) {
+      const seconds = getPlaythroughSecondsElapsed();
+      timerDraw(TimerType.MAIN, seconds);
+    }
   }
 
   drawNoHitTimer(): void {