+++ src/main.ts
@@ -29,6 +29,7 @@ import { PickupRemoval } from "./classes/features/PickupRemoval";
 import { PillRemoval } from "./classes/features/PillRemoval";
 import { PreventEndMegaSatan } from "./classes/features/PreventEndMegaSatan";
 import { PreventPause } from "./classes/features/PreventPause";
+import { PreventSaveAndQuit } from "./classes/features/PreventSaveAndQuit";
 import { PreventVictoryLapPopup } from "./classes/features/PreventVictoryLapPopup";
 import { RemoveDonationMachines } from "./classes/features/RemoveDonationMachines";
 import { RemoveGlitchedCollectibles } from "./classes/features/RemoveGlitchedCollectibles";
@@ -71,6 +72,7 @@ const MOD_FEATURES = [
   PillRemoval,
   PreventEndMegaSatan,
   PreventPause,
+  PreventSaveAndQuit,
   PreventVictoryLapPopup,
   RemoveDonationMachines,
   RemoveGlitchedCollectibles,